Move airflow integration config into vendor package
